requirements.txt
numpy==1.16.4
pandas==0.25.3
matplotlib==3.3.0
opencv-python==4.5.1.48
Pillow==7.2.0
imutils==0.5.4
pyzbar==0.1.8
qrcode==6.1
requests==2.25.1
Flask==1.1.2
Flask-Cors==3.0.10